fixed a bunch of bugs
The sort method now sorts strings lexicographically, instead of alphabetically.
Also, fixed some bugs with the format of the output files of the timing methods.

+++ src/assign04/AreAnagramsCheckAnalysis.java
@@ -42,8 +42,8 @@ public class AreAnagramsCheckAnalysis {
 			fileRead = new Scanner(new File(fileName));
 
 			while(fileRead.hasNextLine()) {
-				int N = fileRead.nextInt();
-				int runTime = fileRead.nextInt();
+				long N = fileRead.nextLong();
+				long runTime = fileRead.nextLong();
 
 				System.out.print(N + "\t|  ");
 				System.out.println(formatter.format(runTime) + "\t" +
